Install ngrok on the E2E test machine with static address
In order to have github runner E2E tests, we need to get hooks for a
github application to the virtual machine on which E2E runs even if we
create a new virtual machine with each run.

Ngrok is a tool creating tunnel to expose local web servers to public
with static address support. Installing it on the virtual machine and
creating a new github app with that static address allows us to get
hooks to virtual machines created by E2E test.