Run ngrok on e2e test machine to expose local server to public with static domain
In order to run github runner e2e tests, local server run on the e2e test
machine must be accessible from the public with a static domain. So, the
github test app we will create can access e2e test machine even if different
hosts can be used.

Ngrok is a tool which exposes local server to public. It also supports static
domains. So, it can be used to access local server with a same domain. That
static domain obtained from ngrok will be used as the base uri for the github
test app will be created. It will allow test app to send hooks to the local
server running on the e2e test machine.